Start to add functionality to the SOS button, mostly in the tap configuration.
Make a new folder for singletons.
Make a class for the SOS button, allowing for the number of clicks to be varied. Completely redesign the code for the SOS button, but it looks the same. This is for the multi-click configuration noted above.
Use the sos_manager singleton to handle the configuration, but it is not yet hooked up to the dropdownmenu. So far, the default value is 3.